rename package as kalelinear
