+++ kalelinear/estimator/_manifold_learn.py
