+++ kalelinear/estimator/base.py
